Created Stateful class. The statefulness functionality has been moved from Module to Stateful. This makes it clearer which classes leverage this functionality.
